Delete the v1 MCP assembly and retire the transitional naming
