+++ apps/cloud/src/mcp/agent-handler.ts
@@ -9,7 +9,7 @@ import {
   type AuthOutcome,
   type McpResource,
 } from "@executor-js/host-mcp";
-import { requestBodyFromRequest } from "@executor-js/host-mcp/tool-server-v2";
+import { requestBodyFromRequest } from "@executor-js/host-mcp/tool-server";
 import {
   currentPropagationHeaders,
   readArtifactsEnabled,
