+++ packages/hosts/mcp-apps-shell/src/shell-resource.smoke.test.ts
@@ -4,7 +4,7 @@ import { Client } from "@modelcontextprotocol/sdk/client/index.js";
 import { InMemoryTransport } from "@modelcontextprotocol/sdk/inMemory.js";
 import { EXTENSION_ID, RESOURCE_MIME_TYPE } from "@modelcontextprotocol/ext-apps/server";
 import type { ClientCapabilities } from "@modelcontextprotocol/sdk/types.js";
-import { createExecutorMcpServer } from "@executor-js/host-mcp/tool-server";
+import { buildMcpServer } from "@executor-js/host-mcp/tool-server";
 import { MCP_APPS_SHELL_RESOURCE_URI } from "@executor-js/host-mcp/create-artifact";
 import type { ExecutionEngine } from "@executor-js/execution";
 
@@ -40,8 +40,12 @@ describe("MCP-Apps shell resource", () => {
     const mcpServer = await Effect.runPromise(
       // Artifacts are opt-in per connection; the shell resource only exists on
       // a session that asked for them.
-      createExecutorMcpServer({
+      buildMcpServer({
         engine: stubEngine,
+        appsEnabled: false,
+        requestStateSigningKey: new Uint8Array(32).fill(17),
+        requestStatePrincipal: "shell-resource-test-principal",
+        sessionful: true,
         loadAppShellHtml: loadMcpAppsShellHtml,
         artifactsEnabled: true,
       }),
