+++ packages/hosts/mcp/package.json
@@ -12,10 +12,6 @@
       "types": "./src/tool-server.ts",
       "default": "./src/tool-server.ts"
     },
-    "./tool-server-v2": {
-      "types": "./src/tool-server-v2.ts",
-      "default": "./src/tool-server-v2.ts"
-    },
     "./mcp-apps": {
       "types": "./src/mcp-apps.ts",
       "default": "./src/mcp-apps.ts"
@@ -52,12 +48,9 @@
     "typecheck:slow": "bunx tsc --noEmit -p tsconfig.json"
   },
   "dependencies": {
-    "@cfworker/json-schema": "^4.1.1",
     "@executor-js/execution": "workspace:*",
     "@executor-js/sdk": "workspace:*",
     "@modelcontextprotocol/core": "2.0.0",
-    "@modelcontextprotocol/ext-apps": "^1.7.4",
-    "@modelcontextprotocol/sdk": "^1.29.0",
     "@modelcontextprotocol/server": "2.0.0",
     "effect": "catalog:",
     "zod": "4.3.6"
@@ -65,6 +58,7 @@
   "devDependencies": {
     "@effect/vitest": "catalog:",
     "@modelcontextprotocol/client": "2.0.0",
+    "@modelcontextprotocol/sdk": "^1.29.0",
     "@types/node": "catalog:",
     "bun-types": "catalog:",
     "vitest": "catalog:"
