+++ packages/hosts/mcp/src/artifacts-tools.test.ts
@@ -1,9 +1,7 @@
 import { describe, expect, it, vi } from "@effect/vitest";
 import { Data, Effect } from "effect";
-import { Client } from "@modelcontextprotocol/sdk/client/index.js";
-import { InMemoryTransport } from "@modelcontextprotocol/sdk/inMemory.js";
-import type { ClientCapabilities } from "@modelcontextprotocol/sdk/types.js";
-import { EXTENSION_ID, RESOURCE_MIME_TYPE } from "@modelcontextprotocol/ext-apps/server";
+import { Client } from "@modelcontextprotocol/client";
+import { InMemoryTransport, type ClientCapabilities } from "@modelcontextprotocol/server";
 import type * as Cause from "effect/Cause";
 
 import {
@@ -18,7 +16,8 @@ import type { ExecutionEngine, ExecutionResult } from "@executor-js/execution";
 import type { BindableConnection } from "./artifact-bindings";
 import { readArtifactsEnabled } from "./browser-approval";
 import { MCP_APPS_SHELL_RESOURCE_URI, artifactUrlFor } from "./create-artifact";
-import { createExecutorMcpServer, type ExecutorMcpServerConfig } from "./tool-server";
+import { EXTENSION_ID, RESOURCE_MIME_TYPE } from "./mcp-apps";
+import { buildMcpServer, type ExecutorMcpServerConfig } from "./tool-server";
 
 /** The caller's connection inventory, as `create-artifact` sees it when it
  *  binds an artifact's integration roles. */
@@ -130,6 +129,14 @@ const makeArtifactStore = () => {
   };
 };
 
+const TEST_REQUEST_STATE_KEY = new Uint8Array(32).fill(11);
+const SESSION_SERVER_OPTIONS = {
+  appsEnabled: false,
+  requestStateSigningKey: TEST_REQUEST_STATE_KEY,
+  requestStatePrincipal: "artifact-test-principal",
+  sessionful: true,
+} as const;
+
 const withClient = async <E extends Cause.YieldableError>(
   engine: ExecutionEngine<E>,
   capabilities: ClientCapabilities,
@@ -137,8 +144,9 @@ const withClient = async <E extends Cause.YieldableError>(
   config?: Partial<ExecutorMcpServerConfig<E>>,
 ) => {
   const mcpServer = await Effect.runPromise(
-    createExecutorMcpServer({
+    buildMcpServer({
       engine,
+      ...SESSION_SERVER_OPTIONS,
       loadAppShellHtml: () => Promise.resolve(SHELL_HTML),
       // Artifacts are on by default and nearly every test in this file
       // exercises the artifact surface; spelled out here anyway so the cases
@@ -164,13 +172,10 @@ const withClient = async <E extends Cause.YieldableError>(
 
 const SHELL_HTML = "<!doctype html><html><body><div id='root'></div></body></html>";
 
-// What a client that renders MCP Apps advertises at `initialize`. The SDK's
-// `ClientCapabilities` has no `extensions` field yet (pending SEP-1724), which
-// is exactly why ext-apps ships `getUiCapability` to read it.
-// oxlint-disable-next-line executor/no-double-cast -- boundary: MCP SDK ClientCapabilities predates the ext-apps `extensions` field
+// What a client that renders MCP Apps advertises at `initialize`.
 const APPS_CAPS = {
   extensions: { [EXTENSION_ID]: { mimeTypes: [RESOURCE_MIME_TYPE] } },
-} as unknown as ClientCapabilities;
+} satisfies ClientCapabilities;
 
 const NO_APPS_CAPS: ClientCapabilities = {};
 
@@ -244,8 +249,9 @@ describe("MCP host — artifact tool visibility", () => {
   it("keeps the app-only tools visible on a cold restore that replays no initialize", async () => {
     const store = makeArtifactStore();
     const mcpServer = await Effect.runPromise(
-      createExecutorMcpServer({
+      buildMcpServer({
         engine: makeStubEngine({}),
+        ...SESSION_SERVER_OPTIONS,
         artifacts: store.port,
         artifactsEnabled: true,
         loadAppShellHtml: () => Promise.resolve(SHELL_HTML),
@@ -307,8 +313,9 @@ describe("MCP host — artifact tool visibility", () => {
   it("renders inline on a restore that replays initialize without the initialized notification", async () => {
     const store = makeArtifactStore();
     const mcpServer = await Effect.runPromise(
-      createExecutorMcpServer({
+      buildMcpServer({
         engine: makeStubEngine({}),
+        ...SESSION_SERVER_OPTIONS,
         artifacts: store.port,
         artifactsEnabled: true,
         loadAppShellHtml: () => Promise.resolve(SHELL_HTML),
@@ -460,8 +467,9 @@ describe("MCP host — artifact tool visibility", () => {
   it("registers no ui tools at all when no shell loader is configured", async () => {
     const store = makeArtifactStore();
     const mcpServer = await Effect.runPromise(
-      createExecutorMcpServer({
+      buildMcpServer({
         engine: makeStubEngine({}),
+        ...SESSION_SERVER_OPTIONS,
         artifacts: store.port,
         // Opted in, so the only thing withholding the surface is the missing
         // shell loader — otherwise this would pass on the connection default.
@@ -1303,7 +1311,7 @@ describe("MCP host — artifact retrieval", () => {
 
         await client.callTool({
           name: "create-artifact",
-          arguments: { code: COUNTER_CODE, title: "Dashboard v2", artifactId: "art_1" },
+          arguments: { code: COUNTER_CODE, title: "Revised dashboard", artifactId: "art_1" },
         });
         expect(usage).toEqual(["created", "updated"]);
