+++ packages/hosts/mcp/src/tool-server.ts
@@ -1,225 +1,310 @@
-import { Data, Effect, Match } from "effect";
+/**
+ * MCP server assembly shared by stateless modern requests and sessionful
+ * connections. Stateless callers supply request-scoped capability policy;
+ * sessionful callers register the full surface once and read negotiated client
+ * capabilities from the live server.
+ */
+import { Data, Effect, Match, Option, Schema } from "effect";
 import * as Cause from "effect/Cause";
-import { Validator } from "@cfworker/json-schema";
 import {
-  getUiCapability,
-  registerAppResource,
-  registerAppTool,
-} from "@modelcontextprotocol/ext-apps/server";
-import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
-import type { ClientCapabilities } from "@modelcontextprotocol/sdk/types.js";
-import type {
-  jsonSchemaValidator,
-  JsonSchemaType,
-  JsonSchemaValidator,
-} from "@modelcontextprotocol/sdk/validation/types.js";
+  acceptedContent,
+  CLIENT_CAPABILITIES_META_KEY,
+  createRequestStateCodec,
+  fromJsonSchema,
+  inputRequired,
+  inputResponse,
+  McpServer,
+  type CallToolResult,
+  type InputRequiredResult,
+  type ServerContext,
+} from "@modelcontextprotocol/server";
 import * as z from "zod/v4";
 
-import type { ElicitationContext, ElicitationHandler, ElicitationRequest } from "@executor-js/sdk";
-import { ElicitationResponse } from "@executor-js/sdk";
+import type { ElicitationRequest } from "@executor-js/sdk";
 
 import {
-  createExecutorMcpServerAssembly,
-  elicitationRequestTag,
+  getUiCapability,
+  EXTENSION_ID,
+  registerAppResource,
+  registerAppTool,
+  RESOURCE_MIME_TYPE,
+  RESOURCE_URI_META_KEY,
+  type McpAppsClientCapabilities,
+  type McpAppToolMeta,
+} from "./mcp-apps";
+import {
+  buildExecutorMcpTools,
   type ExecutorMcpAssembly,
-  type ExecutorMcpServerConfig,
+  type ExecutorMcpToolConfig,
   type McpHandlerResult,
   type McpRequestJoinKeys,
   type McpToolResult,
   type NativeExecutionServices,
-} from "./tool-server-shared";
+} from "./tool-server-core";
 
-export { formatMcpExecutionOutcome, PAUSED_APPROVAL_TIMEOUT_MS } from "./tool-server-shared";
+export { formatMcpExecutionOutcome, PAUSED_APPROVAL_TIMEOUT_MS } from "./tool-server-core";
 export type {
   BrowserApprovalStore,
-  ExecutorMcpServerConfig,
+  ExecutorMcpToolConfig,
   McpArtifactsPort,
   McpConnectionsPort,
   McpToolResult,
   PausedExecutionHooks,
   ResumeFallbackOutcome,
   ResumeUnavailableStatus,
-} from "./tool-server-shared";
+} from "./tool-server-core";
 
-// Workers-compatible JSON Schema validator (replaces Ajv, which uses new Function()).
-class CfWorkerJsonSchemaValidator implements jsonSchemaValidator {
-  getValidator<T>(schema: JsonSchemaType): JsonSchemaValidator<T> {
-    const validator = new Validator(schema as Record<string, unknown>, "2020-12", false);
-    return (input: unknown) => {
-      const result = validator.validate(input);
-      if (result.valid) {
-        return { valid: true, data: input as T, errorMessage: undefined };
-      }
-      const errorMessage = result.errors
-        .map((error) => `${error.instanceLocation}: ${error.error}`)
-        .join("; ");
-      return { valid: false, data: undefined, errorMessage };
-    };
-  }
-}
+const NATIVE_ELICITATION_RESPONSE_KEY = "elicitation";
 
-class McpNativeElicitationTransportError extends Data.TaggedError(
-  "McpNativeElicitationTransportError",
-)<{
-  readonly cause: unknown;
-}> {}
+const NativeRequestStateSchema = Schema.Struct({ executionId: Schema.String });
+/** Verified payload carried by a modern native-elicitation continuation. */
+export type NativeRequestState = typeof NativeRequestStateSchema.Type;
+const decodeNativeRequestState = Schema.decodeUnknownOption(NativeRequestStateSchema);
 
-type ElicitInputParams =
-  | {
-      mode?: "form";
-      message: string;
-      requestedSchema: { readonly [key: string]: unknown };
-    }
-  | { mode: "url"; message: string; url: string; elicitationId: string };
-
-const requestedSchemaIsNonEmpty = (request: ElicitationRequest): boolean =>
-  Match.value(request).pipe(
-    Match.tag("FormElicitation", (form) => Object.keys(form.requestedSchema).length > 0),
-    Match.tag("UrlElicitation", () => false),
-    Match.exhaustive,
-  );
-
-const elicitationRequestUrl = (request: ElicitationRequest): string | undefined =>
-  Match.value(request).pipe(
-    Match.tag("UrlElicitation", (url): string | undefined => url.url),
-    Match.tag("FormElicitation", (): string | undefined => undefined),
-    Match.exhaustive,
-  );
-
-const elicitationRequestToParams: (request: ElicitationRequest) => ElicitInputParams =
-  Match.type<ElicitationRequest>().pipe(
-    Match.tag("UrlElicitation", (url) => ({
-      mode: "url" as const,
-      message: url.message,
-      url: url.url,
-      elicitationId: url.elicitationId,
-    })),
-    Match.tag("FormElicitation", (form) => ({
-      message: form.message,
-      requestedSchema:
-        Object.keys(form.requestedSchema).length === 0
-          ? { type: "object" as const, properties: {} }
-          : form.requestedSchema,
-    })),
-    Match.exhaustive,
-  );
-
-const getElicitationSupport = (server: McpServer): { form: boolean; url: boolean } => {
-  const capabilities = server.server.getClientCapabilities();
-  if (capabilities === undefined || !capabilities.elicitation) return { form: false, url: false };
-  const elicitation = capabilities.elicitation as Record<string, unknown>;
-  return { form: Boolean(elicitation.form), url: Boolean(elicitation.url) };
+type McpServerRequestContext = McpRequestJoinKeys & {
+  readonly serverContext: ServerContext;
 };
 
-const formatBoundaryError = (
-  error: unknown,
-): { name?: string; message: string; stack?: string } => {
-  // oxlint-disable-next-line executor/no-instanceof-error -- boundary: SDK Promise rejection supplies unknown JS errors for debug logging only
-  if (error instanceof Error) {
-    // oxlint-disable-next-line executor/no-unknown-error-message -- boundary: narrowed native Error detail is confined to opt-in debug logging
-    return { name: error.name, message: error.message, stack: error.stack };
-  }
-  // oxlint-disable-next-line executor/no-unknown-error-message -- boundary: fallback log formatting for unknown SDK Promise rejection values
-  return { message: String(error) };
-};
-
-const makeMcpElicitationHandler =
-  (
-    server: McpServer,
-    relatedRequestId: string | number,
-    debugLog: (event: string, data: Record<string, unknown>) => void,
-  ): ElicitationHandler =>
-  (context: ElicitationContext): Effect.Effect<typeof ElicitationResponse.Type> => {
-    const { url: supportsUrl } = getElicitationSupport(server);
-    const params = Match.value(context.request).pipe(
-      Match.tag(
-        "UrlElicitation",
-        (request): ElicitInputParams =>
-          supportsUrl
-            ? elicitationRequestToParams(request)
-            : {
-                message: `${request.message}\n\nPlease visit this URL:\n${request.url}\n\nClick accept once you have completed the flow.`,
-                requestedSchema: { type: "object" as const, properties: {} },
-              },
-      ),
-      Match.tag(
-        "FormElicitation",
-        (request): ElicitInputParams => elicitationRequestToParams(request),
-      ),
-      Match.exhaustive,
-    );
-
-    return Effect.promise(async (): Promise<typeof ElicitationResponse.Type> => {
-      debugLog("elicitation.request", {
-        requestTag: elicitationRequestTag(context.request),
-        supportsUrl,
-        message: context.request.message,
-        hasRequestedSchema: requestedSchemaIsNonEmpty(context.request),
-        url: elicitationRequestUrl(context.request),
-        clientCapabilities: server.server.getClientCapabilities() ?? null,
-      });
-
-      const response = await server.server.elicitInput(
-        params as Parameters<typeof server.server.elicitInput>[0],
-        { relatedRequestId },
-      );
-      debugLog("elicitation.response", {
-        requestTag: elicitationRequestTag(context.request),
-        action: response.action,
-        hasContent:
-          typeof response.content === "object" &&
-          response.content !== null &&
-          Object.keys(response.content).length > 0,
-      });
-      return {
-        action: response.action as typeof ElicitationResponse.Type.action,
-        content: response.content,
-      };
-    }).pipe(
-      Effect.tapDefect((defect) =>
-        Effect.sync(() => {
-          debugLog("elicitation.error", {
-            requestTag: elicitationRequestTag(context.request),
-            error: formatBoundaryError(defect),
-            clientCapabilities: server.server.getClientCapabilities() ?? null,
-          });
-        }),
-      ),
-      Effect.catchDefect((cause) =>
-        // oxlint-disable-next-line executor/no-effect-escape-hatch -- boundary: ElicitationHandler has no error channel, so retain a classified defect for the MCP result boundary
-        Effect.die(new McpNativeElicitationTransportError({ cause })),
-      ),
-    );
+/** Configuration required to build an Executor MCP server. */
+export type ExecutorMcpServerConfig<E extends Cause.YieldableError = Cause.YieldableError> =
+  ExecutorMcpToolConfig<E> & {
+    /** Initial/static MCP Apps policy. Sessionful servers replace it after initialize. */
+    readonly appsEnabled: boolean;
+    /**
+     * Register a connection-lifetime server whose capability-dependent behavior
+     * follows the live initialize-negotiated state. Omitted for stateless modern
+     * request factories, which keep using {@link appsEnabled} as fixed policy.
+     */
+    readonly sessionful?: boolean;
+    /** HMAC key used to sign opaque native-elicitation continuation state. */
+    readonly requestStateSigningKey: Uint8Array | string;
+    /**
+     * Stable identifier of the authenticated principal this server instance
+     * was built for (org/user/subject). Bound into the signed continuation
+     * state so a `requestState` minted for one principal is rejected when
+     * echoed by another — the spec's user-binding MUST for state that
+     * influences authorization. Single-user hosts pass a constant.
+     */
+    readonly requestStatePrincipal: string;
+    /** Lifetime of signed continuation state in seconds; the SDK defaults to ten minutes. */
+    readonly requestStateTtlSeconds?: number;
   };
 
-const requestJoinKeys = (extra: {
-  readonly requestId: string | number;
-  readonly sessionId?: string;
-}): McpRequestJoinKeys => ({
-  requestId: extra.requestId,
-  ...(extra.sessionId === undefined ? {} : { sessionId: extra.sessionId }),
+/** Decide whether client capabilities advertise support for MCP Apps HTML. */
+export const appsEnabledForClientCapabilities = (
+  clientCapabilities: McpAppsClientCapabilities | null | undefined,
+): boolean => Boolean(getUiCapability(clientCapabilities)?.mimeTypes?.includes(RESOURCE_MIME_TYPE));
+
+/** Bind modern continuation state to the ownership identity used by MCP hosts. */
+export const mcpRequestStatePrincipal = (principal: {
+  readonly accountId: string;
+  readonly organizationId: string;
+}): string => `${principal.accountId}\u0000${principal.organizationId}`;
+
+const requestStateBinding = (method: string, principal: string): string =>
+  `${method}\u0000${principal}`;
+
+/** Route-level failure verifying untrusted modern continuation state. */
+export class McpRequestStateVerificationError extends Data.TaggedError(
+  "McpRequestStateVerificationError",
+)<{ readonly cause: unknown }> {}
+
+/**
+ * Verify and parse a modern continuation before a stateless worker uses its
+ * execution id for Durable Object routing.
+ */
+export const verifyNativeRequestState = (input: {
+  readonly state: string;
+  readonly method: string;
+  readonly requestStateSigningKey: Uint8Array | string;
+  readonly requestStatePrincipal: string;
+}): Effect.Effect<NativeRequestState, McpRequestStateVerificationError> => {
+  const codec = createRequestStateCodec<NativeRequestState>({
+    key: input.requestStateSigningKey,
+    bind: () => requestStateBinding(input.method, input.requestStatePrincipal),
+  });
+  return Effect.tryPromise({
+    // The route-level verifier has no handler context. Its codec binding is a
+    // closed value derived from the already-parsed method and principal, so the
+    // SDK callback never observes this inert placeholder.
+    try: async () => {
+      const decoded: unknown = await Reflect.apply(codec.verify, codec, [input.state, null]);
+      return Effect.runPromise(Schema.decodeUnknownEffect(NativeRequestStateSchema)(decoded));
+    },
+    catch: (cause) => new McpRequestStateVerificationError({ cause }),
+  });
+};
+
+const isRecord = (value: unknown): value is Record<string, unknown> =>
+  typeof value === "object" && value !== null && !Array.isArray(value);
+
+const appsClientCapabilitiesFromUnknown = (
+  capabilities: unknown,
+): McpAppsClientCapabilities | null => {
+  if (!isRecord(capabilities)) return null;
+  const extensions = capabilities.extensions;
+  if (!isRecord(extensions)) return null;
+  const ui = extensions[EXTENSION_ID];
+  if (!isRecord(ui)) return null;
+  const mimeTypes = ui.mimeTypes;
+  if (mimeTypes === undefined) return { extensions: { [EXTENSION_ID]: {} } };
+  if (!Array.isArray(mimeTypes) || !mimeTypes.every((value) => typeof value === "string")) {
+    return null;
+  }
+  return { extensions: { [EXTENSION_ID]: { mimeTypes } } };
+};
+
+const elicitationSupportFromUnknown = (
+  capabilities: unknown,
+): { readonly form: boolean; readonly url: boolean } => {
+  if (!isRecord(capabilities) || !isRecord(capabilities.elicitation)) {
+    return { form: false, url: false };
+  }
+  const elicitation = capabilities.elicitation;
+  const hasExplicitModes = "form" in elicitation || "url" in elicitation;
+  return {
+    form: hasExplicitModes ? Boolean(elicitation.form) : true,
+    url: Boolean(elicitation.url),
+  };
+};
+
+/** Parse the MCP Apps capability subset from an already-decoded modern body. */
+export const clientCapabilitiesFromRequestBody = (
+  body: unknown,
+): McpAppsClientCapabilities | null => {
+  if (!isRecord(body)) return null;
+  const params = body.params;
+  if (!isRecord(params)) return null;
+  const metadata = params._meta;
+  if (!isRecord(metadata)) return null;
+  const capabilities = metadata[CLIENT_CAPABILITIES_META_KEY];
+  return appsClientCapabilitiesFromUnknown(capabilities);
+};
+
+/** Parse a cloned HTTP request body without consuming the request itself. */
+export const requestBodyFromRequest = (request: Request): Effect.Effect<unknown> =>
+  Effect.tryPromise({
+    try: () => request.clone().json(),
+    catch: () => null,
+  }).pipe(
+    Effect.match({
+      onFailure: () => null,
+      onSuccess: (body) => body,
+    }),
+  );
+
+/**
+ * Parse the MCP Apps capability subset from a modern request's `_meta`
+ * envelope without consuming the request body used by the SDK handler.
+ */
+export const clientCapabilitiesFromRequest = (
+  request: Request,
+): Effect.Effect<McpAppsClientCapabilities | null> =>
+  requestBodyFromRequest(request).pipe(Effect.map(clientCapabilitiesFromRequestBody));
+
+const requestJoinKeys = (context: ServerContext): McpServerRequestContext => ({
+  requestId: context.mcpReq.id,
+  ...(context.sessionId === undefined ? {} : { sessionId: context.sessionId }),
+  serverContext: context,
 });
 
-const v1ToolResult = (result: McpHandlerResult): McpToolResult =>
-  "resultType" in result
-    ? {
-        content: [{ type: "text", text: "Input-required results are unavailable on SDK v1." }],
-        isError: true,
-      }
-    : result;
+const appToolMeta = (metadata: Record<string, unknown>): McpAppToolMeta | undefined => {
+  const ui = metadata.ui;
+  if (!isRecord(ui)) return undefined;
+  const resourceUri = typeof ui.resourceUri === "string" ? ui.resourceUri : undefined;
+  const visibility = Array.isArray(ui.visibility)
+    ? ui.visibility.filter(
+        (value): value is "model" | "app" => value === "model" || value === "app",
+      )
+    : undefined;
+  return {
+    ...(resourceUri === undefined ? {} : { resourceUri }),
+    ...(visibility === undefined ? {} : { visibility }),
+  };
+};
 
-const createV1Assembly = <E extends Cause.YieldableError>(
+const normalizedAppMetadata = (metadata: Record<string, unknown>) => {
+  const ui = appToolMeta(metadata);
+  const legacyResourceUri = metadata[RESOURCE_URI_META_KEY];
+  return {
+    ...metadata,
+    ...(ui === undefined ? {} : { ui }),
+    ...(typeof legacyResourceUri === "string"
+      ? { [RESOURCE_URI_META_KEY]: legacyResourceUri }
+      : {}),
+  };
+};
+
+const withoutAppMetadata = (metadata: Record<string, unknown>): Record<string, unknown> => {
+  const { ui: _ui, [RESOURCE_URI_META_KEY]: _resourceUri, ...rest } = metadata;
+  return rest;
+};
+
+const visibilityIncludes = (
+  metadata: Record<string, unknown>,
+  visibility: "model" | "app",
+): boolean => appToolMeta(metadata)?.visibility?.includes(visibility) ?? true;
+
+const toolResult = (result: McpHandlerResult): CallToolResult | InputRequiredResult => result;
+
+const elicitationInputRequest = (request: ElicitationRequest) =>
+  Match.value(request).pipe(
+    Match.tag("FormElicitation", (form) =>
+      inputRequired.elicit({
+        message: form.message,
+        requestedSchema:
+          Object.keys(form.requestedSchema).length === 0
+            ? fromJsonSchema({ type: "object" as const, properties: {} })
+            : fromJsonSchema(form.requestedSchema),
+      }),
+    ),
+    Match.tag("UrlElicitation", (url) =>
+      inputRequired.elicitUrl({ message: url.message, url: url.url }),
+    ),
+    Match.exhaustive,
+  );
+
+const missingNativeExecution = (executionId: string): McpToolResult => ({
+  content: [
+    {
+      type: "text",
+      text: `Paused execution is unknown: ${executionId}. Run execute again to start a fresh flow.`,
+    },
+  ],
+  structuredContent: {
+    status: "execution_not_found",
+    executionId,
+  },
+  isError: true,
+});
+
+const createMcpAssembly = <E extends Cause.YieldableError>(
   config: ExecutorMcpServerConfig<E>,
-): ExecutorMcpAssembly<McpServer, McpRequestJoinKeys> => {
+): ExecutorMcpAssembly<McpServer, McpServerRequestContext> => {
+  const sessionful = config.sessionful ?? false;
+  const initialAppsEnabled = sessionful
+    ? (config.restoredAppsEnabled ?? config.appsEnabled)
+    : config.appsEnabled;
+  const requestStateCodec = createRequestStateCodec<NativeRequestState>({
+    key: config.requestStateSigningKey,
+    ...(config.requestStateTtlSeconds === undefined
+      ? {}
+      : { ttlSeconds: config.requestStateTtlSeconds }),
+    bind: (context) => requestStateBinding(context.mcpReq.method, config.requestStatePrincipal),
+  });
+  const verifyRequestState = async (state: string, context: ServerContext) => {
+    const decoded = await requestStateCodec.verify(state, context);
+    return Effect.runPromise(Schema.decodeUnknownEffect(NativeRequestStateSchema)(decoded));
+  };
   const server = new McpServer(
     { name: "executor", version: "1.0.0" },
     {
       capabilities: { resources: {}, tools: {} },
-      jsonSchemaValidator: new CfWorkerJsonSchemaValidator(),
+      requestState: { verify: verifyRequestState },
     },
   );
 
-  const registerTool: ExecutorMcpAssembly<McpServer, McpRequestJoinKeys>["registerTool"] = (
+  const registerTool: ExecutorMcpAssembly<McpServer, McpServerRequestContext>["registerTool"] = (
     name,
     toolConfig,
     callback,
@@ -228,69 +313,151 @@ const createV1Assembly = <E extends Cause.YieldableError>(
     return server.registerTool<z.ZodObject<z.ZodRawShape>, typeof inputSchema>(
       name,
       { ...toolConfig, inputSchema },
-      async (args, extra) => v1ToolResult(await callback(args, requestJoinKeys(extra))),
+      async (args, context) => toolResult(await callback(args, requestJoinKeys(context))),
     );
   };
 
-  const registerApp: ExecutorMcpAssembly<McpServer, McpRequestJoinKeys>["registerAppTool"] = (
+  const registerApp: ExecutorMcpAssembly<McpServer, McpServerRequestContext>["registerAppTool"] = (
     name,
     toolConfig,
     callback,
   ) => {
     const inputSchema = z.object(toolConfig.inputSchema);
-    return registerAppTool<z.ZodObject<z.ZodRawShape>, typeof inputSchema>(
+    const metadata = normalizedAppMetadata(toolConfig._meta);
+    if (!sessionful && !config.appsEnabled && visibilityIncludes(metadata, "model")) {
+      const plainMetadata = withoutAppMetadata(metadata);
+      return server.registerTool<z.ZodObject<z.ZodRawShape>, typeof inputSchema>(
+        name,
+        {
+          ...toolConfig,
+          inputSchema,
+          ...(Object.keys(plainMetadata).length === 0
+            ? { _meta: undefined }
+            : { _meta: plainMetadata }),
+        },
+        async (args, context) => toolResult(await callback(args, requestJoinKeys(context))),
+      );
+    }
+
+    return registerAppTool<typeof inputSchema, z.ZodObject<z.ZodRawShape>>(
       server,
       name,
-      { ...toolConfig, inputSchema },
-      async (args, extra) => v1ToolResult(await callback(args, requestJoinKeys(extra))),
+      { ...toolConfig, inputSchema, _meta: metadata },
+      async (args, context) => toolResult(await callback(args, requestJoinKeys(context))),
     );
+  };
+
+  const nativeInputRequired = async <NativeE extends Cause.YieldableError>(
+    services: NativeExecutionServices<NativeE, McpServerRequestContext>,
+    execution: Parameters<
+      NativeExecutionServices<NativeE, McpServerRequestContext>["executionPaused"]
+    >[0],
+  ): Promise<InputRequiredResult> => {
+    const requestState = await requestStateCodec.mint(
+      { executionId: execution.id },
+      services.requestContext.serverContext,
+    );
+    return inputRequired({
+      inputRequests: {
+        [NATIVE_ELICITATION_RESPONSE_KEY]: elicitationInputRequest(
+          execution.elicitationContext.request,
+        ),
+      },
+      requestState,
+    });
   };
 
   return {
     server,
-    era: "v1",
-    initialAppsEnabled: config.restoredAppsEnabled ?? false,
-    getClientCapabilities: () => server.server.getClientCapabilities() ?? null,
-    getElicitationSupport: () => getElicitationSupport(server),
+    initialAppsEnabled,
+    getClientCapabilities: () =>
+      sessionful ? (server.server.getClientCapabilities() ?? null) : null,
+    getElicitationSupport: () =>
+      sessionful
+        ? elicitationSupportFromUnknown(server.server.getClientCapabilities())
+        : { form: true, url: true },
     getUiCapability: () =>
-      getUiCapability(
-        server.server.getClientCapabilities() as
-          | (ClientCapabilities & { extensions?: Record<string, unknown> })
-          | null,
-      ),
+      sessionful
+        ? getUiCapability(appsClientCapabilitiesFromUnknown(server.server.getClientCapabilities()))
+        : config.appsEnabled
+          ? { mimeTypes: [RESOURCE_MIME_TYPE] }
+          : undefined,
     onInitialized: (callback) => {
-      server.server.oninitialized = callback;
+      if (sessionful) server.server.oninitialized = callback;
     },
     registerTool,
     registerAppTool: registerApp,
     registerAppResource: (name, uri, resourceConfig, callback) => {
+      if (!sessionful && !config.appsEnabled) return;
       registerAppResource(server, name, uri, resourceConfig, async () => {
         const result = await callback();
         return { contents: [...result.contents] };
       });
     },
     executeNative: <NativeE extends Cause.YieldableError>(
-      services: NativeExecutionServices<NativeE, McpRequestJoinKeys>,
+      services: NativeExecutionServices<NativeE, McpServerRequestContext>,
     ) =>
-      services.engine
-        .execute(services.code, {
-          onElicitation: makeMcpElicitationHandler(
-            server,
-            services.requestContext.requestId,
-            services.debugLog,
-          ),
-        })
-        .pipe(Effect.map(services.complete)),
+      Effect.gen(function* () {
+        const decodedState = decodeNativeRequestState(
+          services.requestContext.serverContext.mcpReq.requestState<unknown>(),
+        );
+
+        if (Option.isSome(decodedState)) {
+          const paused = yield* services.engine.getPausedExecution(decodedState.value.executionId);
+          if (!paused) return missingNativeExecution(decodedState.value.executionId);
+          const response = inputResponse(
+            services.requestContext.serverContext.mcpReq.inputResponses,
+            NATIVE_ELICITATION_RESPONSE_KEY,
+          );
+          if (response.kind === "elicit") {
+            const content = Match.value(paused.elicitationContext.request).pipe(
+              Match.tag("UrlElicitation", () => response.content),
+              Match.tag("FormElicitation", (form) =>
+                response.action === "accept"
+                  ? acceptedContent(
+                      services.requestContext.serverContext.mcpReq.inputResponses,
+                      NATIVE_ELICITATION_RESPONSE_KEY,
+                      fromJsonSchema<Record<string, unknown>>(
+                        Object.keys(form.requestedSchema).length === 0
+                          ? { type: "object", properties: {} }
+                          : form.requestedSchema,
+                      ),
+                    )
+                  : response.content,
+              ),
+              Match.exhaustive,
+            );
+            if (response.action === "accept" && content === undefined) {
+              return yield* Effect.promise(() => nativeInputRequired(services, paused));
+            }
+            const outcome = yield* services.resume(decodedState.value.executionId, {
+              action: response.action,
+              content,
+            });
+            if (!outcome) return missingNativeExecution(decodedState.value.executionId);
+            if (outcome.status === "completed") return services.complete(outcome.result);
+            yield* services.executionPaused(outcome.execution);
+            return yield* Effect.promise(() => nativeInputRequired(services, outcome.execution));
+          }
+
+          return yield* Effect.promise(() => nativeInputRequired(services, paused));
+        }
+
+        const outcome = yield* services.engine.executeWithPause(services.code);
+        if (outcome.status === "completed") return services.complete(outcome.result);
+        yield* services.executionPaused(outcome.execution);
+        return yield* Effect.promise(() => nativeInputRequired(services, outcome.execution));
+      }),
   };
 };
 
 /**
- * Build the legacy SDK v1 Executor MCP tool server.
+ * Build one Executor MCP server.
  *
- * Its public signature and wire behavior remain unchanged; SDK-specific
- * construction and native elicitation are confined to this assembly.
+ * Stateless hosts must reuse the signing key across every request that can
+ * participate in the same native-elicitation continuation flow. Sessionful
+ * hosts keep one instance connected and may use a connection-lifetime key.
  */
-export const createExecutorMcpServer = <E extends Cause.YieldableError>(
+export const buildMcpServer = <E extends Cause.YieldableError>(
   config: ExecutorMcpServerConfig<E>,
-): Effect.Effect<McpServer> =>
-  createExecutorMcpServerAssembly(config, () => createV1Assembly(config));
+): Effect.Effect<McpServer> => buildExecutorMcpTools(config, () => createMcpAssembly(config));
